requirements.txt
python-docx
docx2txt
PyPDF2
requests
python-dotenv